Move profiles and locations in profiles app
