+++ boilerplate/apps/profiles.app.js
@@ -1,0 +1,8 @@
+const { ProfilesApp } = require('@activitypods/profiles');
+
+module.exports = {
+  mixins: [ProfilesApp],
+  settings: {
+    publicProfile: true
+  }
+};
