+++ boilerplate/package.json
@@ -15,6 +15,7 @@
     "@activitypods/events": "1.1.0",
     "@activitypods/marketplace": "1.1.0",
     "@activitypods/messages": "1.1.0",
+    "@activitypods/profiles": "1.1.0",
     "@activitypods/synchronizer": "1.1.0",
     "@semapps/notifications": "0.4.0-alpha.32",
     "dotenv-flow": "^3.2.0",
