+++ packages/contacts/app.js
@@ -1,13 +1,10 @@
-const LocationService = require('./services/location');
-const ProfileService = require('./services/profile');
+const ManagerService = require('./services/manager');
 const RequestService = require('./services/request');
 
 const ContactsApp = {
   name: 'contacts',
   created() {
-    this.broker.createService(ProfileService);
-
-    this.broker.createService(LocationService);
+    this.broker.createService(ManagerService);
 
     this.broker.createService(RequestService);
   },
