+++ packages/profiles/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ProfilesApp: require('./app'),
+};
