+++ packages/profiles/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@activitypods/profiles",
+  "version": "1.1.0",
+  "description": "ActivityPods app for users' profiles and locations",
+  "license": "Apache-2.0",
+  "author": "Virtual Assembly",
+  "dependencies": {
+    "@activitypods/synchronizer": "1.1.0",
+    "@semapps/activitypub": "0.4.0-alpha.36",
+    "@semapps/ldp": "0.4.0-alpha.36",
+    "@semapps/mime-types": "0.4.0-alpha.36"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
