+++ packages/profiles/services/location.js
@@ -1,7 +1,7 @@
 const { ControlledContainerMixin } = require('@semapps/ldp');
 
 module.exports = {
-  name: 'contacts.location',
+  name: 'profiles.location',
   mixins: [ControlledContainerMixin],
   settings: {
     path: '/locations',
