+++ packages/profiles/services/profile.js
@@ -1,13 +1,14 @@
 const { ControlledContainerMixin } = require('@semapps/ldp');
-const { OBJECT_TYPES, ActivitiesHandlerMixin } = require('@semapps/activitypub');
+const { OBJECT_TYPES } = require('@semapps/activitypub');
 const { MIME_TYPES } = require('@semapps/mime-types');
 const { SynchronizerMixin } = require('@activitypods/synchronizer');
-const { REMOVE_CONTACT } = require("../config/patterns");
 
 module.exports = {
-  name: 'contacts.profile',
-  mixins: [SynchronizerMixin, ControlledContainerMixin, ActivitiesHandlerMixin],
+  name: 'profiles.profile',
+  mixins: [SynchronizerMixin, ControlledContainerMixin],
   settings: {
+    publicProfile: false,
+    // ControlledContainerMixin settings
     path: '/profiles',
     acceptedTypes: ['vcard:Individual', OBJECT_TYPES.PROFILE],
     permissions: {},
@@ -23,6 +24,7 @@ module.exports = {
 
       const profileUri = await this.actions.post({
         containerUri,
+        slug: 'me',
         resource: {
           '@type': ['vcard:Individual', OBJECT_TYPES.PROFILE],
           'vcard:fn': profileData.familyName
@@ -36,6 +38,18 @@ module.exports = {
         webId,
       });
 
+      if (this.settings.publicProfile) {
+        await ctx.call('webacl.resource.addRights', {
+          resourceUri: profileUri,
+          additionalRights: {
+            anon: {
+              read: true,
+            }
+          },
+          webId,
+        });
+      }
+
       await ctx.call('ldp.resource.patch', {
         resource: {
           '@id': webId,
@@ -45,6 +59,7 @@ module.exports = {
         webId,
       });
 
+      // TODO put this on the contacts app
       // Create a WebACL group for the user's contact
       const { groupUri: contactsGroupUri } = await ctx.call('webacl.group.create', {
         groupSlug: new URL(webId).pathname + '/contacts',
@@ -62,30 +77,6 @@ module.exports = {
         },
         webId,
       });
-    },
-  },
-  activities: {
-    removeContact: {
-      match: REMOVE_CONTACT,
-      async onEmit(ctx, activity, emitterUri) {
-        if (!activity.origin)
-          throw new Error('The origin property is missing from the Remove activity');
-
-        if (!activity.origin.startsWith(emitterUri))
-          throw new Error(`Cannot remove from collection ${activity.origin} as it is not owned by the emitter`);
-
-        await ctx.call('activitypub.collection.detach', {
-          collectionUri: activity.origin,
-          item: activity.object.id,
-        });
-
-        const actor = await ctx.call('activitypub.actor.get', { actorUri: activity.object.id, webId: activity.object.id });
-
-        await ctx.call('activitypub.object.deleteFromCache', {
-          actorUri: emitterUri,
-          objectUri: actor.url,
-        });
-      }
     },
   }
 };
